% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.np-package}
\alias{mlr3learners.np}
\alias{mlr3learners.np-package}
\title{mlr3learners.np: Learners from the {np} package for 'mlr3'}
\description{
Adds learners from the {np} package to 'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
